调整Example中几个api的调用
