+++ example/main/java/cn/jmessage/api/examples/MessageExample.java
@@ -15,7 +15,7 @@ public class MessageExample {
     private static final String appkey = "242780bfdd7315dc1989fe2b";
     private static final String masterSecret = "2f5ced2bef64167950e63d13";
 
-    public static void testGetUserInfo() {
+    public static void testSendSingleTextByAdmin() {
         JMessageClient client = new JMessageClient(appkey, masterSecret);
 
         try {
